created EmptyClusterPanelMatchCorrectnessTest created PanelMatchSimulator
