+++ src/main/kotlin/org/wfanet/measurement/loadtest/panelmatch/BUILD.bazel
@@ -1,0 +1,30 @@
+load("@wfa_common_jvm//build/rules_kotlin:defs.bzl", "kt_jvm_library")
+
+package(
+    default_testonly = True,
+    default_visibility = [
+        "//src/test/kotlin/org/wfanet/panelmatch/integration:__subpackages__",
+    ],
+)
+
+kt_jvm_library(
+    name = "simulator",
+    srcs = ["PanelMatchSimulator.kt"],
+    deps = [
+        "//src/main/proto/wfa/measurement/api/v2alpha:exchange_java_proto",
+        "//src/main/proto/wfa/measurement/api/v2alpha:exchange_workflow_kt_jvm_proto",
+        "//src/main/proto/wfa/measurement/api/v2alpha:exchanges_service_kt_jvm_grpc",
+        "//src/main/proto/wfa/measurement/api/v2alpha:exchanges_service_kt_jvm_proto",
+        "//src/main/proto/wfa/measurement/api/v2alpha:exchange_step_attempt_kt_jvm_proto",
+        "//src/main/kotlin/org/wfanet/measurement/loadtest/panelmatchresourcesetup",
+        "//src/main/kotlin/org/wfanet/measurement/loadtest/resourcesetup:resource_setup",
+        "//src/main/kotlin/org/wfanet/panelmatch/client/common:common",
+        "//src/main/kotlin/org/wfanet/panelmatch/client/deploy",
+        "//src/main/proto/wfa/panelmatch/client/storage:storage_details_java_proto",
+        "//src/main/proto/wfa/panelmatch/client/storage:storage_details_kt_jvm_proto",
+        "@wfa_common_jvm//imports/kotlin/com/google/protobuf/kotlin",
+        "@wfa_common_jvm//imports/java/com/google/common/truth",
+        "@wfa_common_jvm//imports/java/com/google/common/truth/extensions/proto",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/storage/forwarded",
+    ],
+)
