+++ src/main/kotlin/org/wfanet/panelmatch/client/deploy/BUILD.bazel
@@ -4,6 +4,7 @@ load("//build:panel_exchange_client_maven_deps.bzl", "panel_exchange_client_mave
 
 package(default_visibility = [
     "//src/main/kotlin/org/wfanet/panelmatch:__subpackages__",
+    "//src/main/kotlin/org/wfanet/measurement/loadtest/panelmatch:__subpackages__",
     "//src/test/kotlin/org/wfanet/panelmatch:__subpackages__",
 ])
